requirements.txt
# Core Dependencies - Always Required
flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
werkzeug==3.0.1
requests==2.31.0
numpy==1.24.3

# Optional Dependencies for Transformers Mode
# Install these if you don't have Ollama or want to use local models
torch==2.1.0
transformers==4.36.0
accelerate==0.25.0
sentencepiece==0.1.99
bitsandbytes==0.41.3
scipy==1.11.4
safetensors==0.4.1

# Development Dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
